+++ components/src/Radio/RadioGroup.js
@@ -68,7 +68,7 @@ BaseRadioGroup.propTypes = {
   onChange: PropTypes.func,
   className: PropTypes.string,
   id: PropTypes.string,
-  helpText: PropTypes.string,
+  helpText: PropTypes.node,
   requirementText: PropTypes.string
 };