+++ docs/src/shared/inputProps.js
@@ -43,7 +43,7 @@ const inputProps = [
   },
   {
     name: "helpText",
-    type: "String",
+    type: "Node",
     defaultValue: "null",
     description:
       "Placed below the label to provide assistance on how to fill out a field or the expected format. It can also provide an explanation of why the information is needed and how it will be used."